Do not hide speadsheet statusbar by default
Status bar is crucial for productivity tool (specially when it comes
to office suites and in specific spreadsheets).

Context, this is a regression from:
- https://github.com/nextcloud/richdocuments/pull/2059
- fabed162ea07ac27c031f288f0fab6f027badf78

Even if,  and as noted in
https://github.com/nextcloud/richdocuments/pull/2053#issuecomment-1044371555,
the status bar was still being hidden for safari I think it's best to
don't break the spreadsheet experience for every browser. On top of
that I think the richdocuments frame suffer some changes and that
safari issue might not be present anymore.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>